+++ src/components/chat/login.tsx
@@ -1,7 +1,6 @@
 import { toast } from 'react-toastify'
 import useWallet from '@/hooks/wallets/useWallet'
-import { loginWithCometChat, signUpWithCometChat, joinGroup } from '../../services/chat'
-import { useEffect } from 'react'
+import { loginWithCometChat, signUpWithCometChat } from '../../services/chat'
 import useSafeAddress from '@/hooks/useSafeAddress'
 
 //@ts-ignore
@@ -52,36 +51,10 @@ const Login = ({ setCurrentUser }) => {
     )
   }
 
-  const handleJoin = async () => {
-    await toast.promise(
-      new Promise(async (resolve, reject) => {
-        await joinGroup(safeAddress!)
-          .then((user) => {
-            console.log(user)
-            resolve(user)
-          })
-          .catch((err) => {
-            console.log(err)
-            reject(err)
-          })
-      }),
-      {
-        pending: 'Signing up...',
-        success: 'Signned up successful 👌',
-        error: 'Error, maybe you should login instead? 🤯',
-      },
-    )
-  }
-
-  useEffect(() => {
-    handleLogin()
-  }, [wallet])
-
   return (
     <div className="flex justify-start items-center space-x-2">
       <button onClick={handleLogin}>Login Now</button>
       <button onClick={handleSignup}>Signup Now</button>
-      <button onClick={handleJoin}>Signup Now</button>
     </div>
   )
 }
